LLMOps/requirements.txt
python-dotenv
google-cloud-aiplatform
vertexai==1.71.1
pandas==2.3.1
pyarrow==21.0.0
scikit-learn==1.7.1